+++ sumo/cli/bandplot.py
@@ -60,6 +60,8 @@ def bandplot(
     scissor=None,
     ylabel="Energy (eV)",
     dos_label=None,
+    zero_line=False,
+    zero_energy=None,
     elements=None,
     lm_orbitals=None,
     atoms=None,
@@ -167,9 +169,14 @@ def bandplot(
             ``False`` (fractional coordinates).
         scissor (:obj:`float`, optional): Apply a scissor operator (rigid shift
             of the CBM), use with caution if applying to metals.
-        dos_file (:obj:'str', optional): Path to vasprun.xml file from which to
+        dos_file (:obj:`str`, optional): Path to vasprun.xml file from which to
             read the density of states information. If set, the density of
             states will be plotted alongside the bandstructure.
+        zero_line (:obj:`bool`, optional): If true, draw a horizontal line at
+            zero energy. (To adjust where zero energy sits, use zero_energy.)
+        zero_energy (:obj:`float`, optional): Energy offset determining position
+            of zero energy. By default, this is the VBM. It may be useful to
+            set this to e.g. a calculated self-consistent Fermi energy.
         elements (:obj:`dict`, optional): The elements and orbitals to extract
             from the projected density of states. Should be provided as a
             :obj:`dict` with the keys as the element names and corresponding
@@ -394,6 +401,8 @@ def bandplot(
             interpolate_factor=interpolate_factor,
             circle_size=circle_size,
             zero_to_efermi=True,
+            zero_line=zero_line,
+            zero_energy=zero_energy,
             ymin=ymin,
             ymax=ymax,
             height=height,
@@ -412,6 +421,8 @@ def bandplot(
     else:
         plt = plotter.get_plot(
             zero_to_efermi=True,
+            zero_line=zero_line,
+            zero_energy=zero_energy,
             ymin=ymin,
             ymax=ymax,
             height=height,
@@ -656,6 +667,20 @@ def _get_parser():
     parser.add_argument(
         "--dos", default=None, help="path to density of states vasprun.xml"
     )
+
+    parser.add_argument(
+        "--zero-line",
+        action="store_true", dest="zero_line",
+        help="Plot horizontal line at energy zero")
+
+    parser.add_argument(
+        "--zero-energy",
+        type=float,
+        dest="zero_energy",
+        default=None,
+        help=("Reference energy: energy will be shifted to place this energy "
+              "at zero. If not specified, zero will be set to the VBM.")
+    )
     parser.add_argument(
         "--elements",
         type=_el_orb,
@@ -810,6 +835,8 @@ def main():
         ylabel=args.ylabel,
         dos_label=args.dos_label,
         dos_file=args.dos,
+        zero_line=args.zero_line,
+        zero_energy=args.zero_energy,
         elements=args.elements,
         lm_orbitals=args.orbitals,
         atoms=args.atoms,