+++ sumo/plotting/dos_plotter.py
@@ -15,6 +15,7 @@ from pymatgen.electronic_structure.core import Spin
 from sumo.electronic_structure.dos import sort_orbitals
 from sumo.plotting import (
     colour_cache,
+    draw_themed_line,
     pretty_plot,
     pretty_subplot,
     styled_plot,
@@ -70,6 +71,7 @@ class SDOSPlotter:
         legend_cutoff=3,
         subplot=False,
         zero_to_efermi=True,
+        zero_energy=None,
         cache=None,
         spin=None,
     ):
@@ -103,6 +105,9 @@ class SDOSPlotter:
                 each element on separate subplots. Defaults to ``False``.
             zero_to_efermi (:obj:`bool`, optional): Normalise the plot such
                 that the Fermi level is set as 0 eV.
+            zero_energy (:obj:`float`, optional): If provided, this value is
+                used as the 0 eV reference. Otherwise, behaviour depends on
+                zero_to_efermi.
             cache (:obj:`dict`, optional): Cache object tracking how colours
                 have been assigned to orbitals. The format is the same as the
                 "colours" dict. This defaults to the module-level
@@ -146,10 +151,19 @@ class SDOSPlotter:
         if cache is None:
             cache = colour_cache
 
-        # mask needed to prevent unwanted data in pdf and for finding y limit
         dos = self._dos
         pdos = self._pdos
-        eners = dos.energies - dos.efermi if zero_to_efermi else dos.energies
+
+        if zero_energy is None:
+            energy_shift = -dos.efermi if zero_to_efermi else 0.
+        elif zero_to_efermi:
+            raise ValueError("Cannot use zero_energy and "
+                             "zero_to_efermi simultaneously.")
+        else:
+            energy_shift = -zero_energy
+        eners = dos.energies + energy_shift
+
+        # mask needed to prevent unwanted data in pdf and for finding y limit
         mask = (eners >= xmin - 0.05) & (eners <= xmax + 0.05)
         plot_data = {"mask": mask, "energies": eners}
         spins = dos.densities.keys()
@@ -236,6 +250,8 @@ class SDOSPlotter:
         xlabel="Energy (eV)",
         ylabel="Arb. units",
         zero_to_efermi=True,
+        zero_line=False,
+        zero_energy=None,
         dpi=400,
         fonts=None,
         plt=None,
@@ -281,6 +297,10 @@ class SDOSPlotter:
             ylabel (:obj:`str`, optional): Label/units for y-axis (i.e. DOS)
             zero_to_efermi (:obj:`bool`, optional): Normalise the plot such
                 that the Fermi level is set as 0 eV.
+            zero_line (:obj:`bool`, optional): Draw a line at 0 eV.
+            zero_energy (:obj:`float`, optional): If provided, this value is
+                used as the 0 eV reference. Otherwise, behaviour depends on
+                zero_to_efermi.
             dpi (:obj:`int`, optional): The dots-per-inch (pixel density) for
                 the image.
             fonts (:obj:`list`, optional): Fonts to use in the plot. Can be a
@@ -294,9 +314,9 @@ class SDOSPlotter:
             no_base_style (:obj:`bool`, optional): Prevent use of sumo base
                 style. This can make alternative styles behave more
                 predictably.
-            spin (:obj:`Spin`, optional): Plot a spin-polarised density of states,
-            "up" or "1" for spin up only, "down" or "-1" for spin down only.
-            Defaults to ``None``.
+            spin (:obj:`Spin`, optional): Plot a spin-polarised density of
+                states, "up" or "1" for spin up only, "down" or "-1" for spin
+                down only.  Defaults to ``None``.
 
         Returns:
             :obj:`matplotlib.pyplot`: The density of states plot.
@@ -310,6 +330,7 @@ class SDOSPlotter:
             legend_cutoff=legend_cutoff,
             subplot=subplot,
             zero_to_efermi=zero_to_efermi,
+            zero_energy=zero_energy,
             spin=spin,
         )
 
@@ -357,6 +378,10 @@ class SDOSPlotter:
                 )
                 ax.plot(energies, densities, label=label, color=line["colour"])
 
+            if zero_line:
+                draw_themed_line(0, ax, orientation='vertical')
+
+            ax.set_ylim(plot_data['ymin'], plot_data['ymax'])
             ax.set_xlim(xmin, xmax)
             if len(spins) == 1:
                 ax.set_ylim(0, plot_data["ymax"])